Remove Makefile template
Summary:
Eliminate the Makefile template from the project. We can just run the script directly from arc init.

I would suggest that if we have new additions of this type to add in future, we either put them into arc init (for stuff which needs to go everywhere) or build it into the maven/gradle/npm build scripts.

This leaves a small job to go around and clean out the Makefiles in various projects.

Test Plan: N/A

Reviewers: O5 elhub/devtools, sergei.prodanov

Reviewed By: O5 elhub/devtools, sergei.prodanov

Subscribers: sergei.prodanov, teamcity

Differential Revision: https://phabricator.elhub.cloud/D1929